#1220 - not hiding window border when it is not resizable anymore, using BorderThickness property instead if want to hide border - added CornerRadius property - can now drag from border too (if BorderThickness is big) - some iconbar fixes - made bootstrap button visible again to show/hide iconbar

+++ Client/ZUI/FloatingWindow/Source/FloatingWindow/FloatingWindowHost.cs
@@ -593,7 +593,7 @@ namespace SilverFlow.Controls
         /// <value>The control displayed in the BottomBar. The default is null.</value>
         public object Bar
         {
-            get { return (double)GetValue(BarProperty); }
+            get { return GetValue(BarProperty); }
             set { SetValue(BarProperty, value); }
         }
